move toggle button to the right menu
