+++ editor/src/components/canvas/right-menu.tsx
@@ -9,6 +9,7 @@ import {
   UtopiaTheme,
   IcnProps,
   LargerIcons,
+  MenuIcons,
 } from 'uuiui'
 import { betterReactMemo, Utils } from 'uuiui-deps'
 import { isLiveMode } from '../editor/editor-modes'
@@ -16,6 +17,8 @@ import { useEditorState } from '../editor/store/store-hook'
 import * as EditorActions from '../editor/actions/action-creators'
 import { EditorAction } from '../editor/action-types'
 import CanvasActions from './canvas-actions'
+import { LeftPaneDefaultWidth } from '../navigator/left-pane'
+import { CanvasVector } from '../../core/shared/math-utils'
 
 export const enum RightMenuTab {
   Insert = 'insert',
@@ -116,6 +119,11 @@ export const RightMenu = betterReactMemo('RightMenu', (props: RightMenuProps) =>
     'RightMenu rightMenuSelectedTab',
   )
 
+  const navigatorVisible = useEditorState(
+    (store) => store.editor.navigator.visible,
+    'RightMenu rightMenuSelectedTab',
+  )
+
   const isInsertMenuSelected = rightMenuSelectedTab === RightMenuTab.Insert
   const isInspectorSelected = rightMenuSelectedTab === RightMenuTab.Inspector
 
@@ -183,6 +191,14 @@ export const RightMenu = betterReactMemo('RightMenu', (props: RightMenuProps) =>
   )
 
   const zoom100pct = React.useCallback(() => dispatch([CanvasActions.zoom(1)]), [dispatch])
+
+  const onClickNavigateTab = React.useCallback(() => {
+    const offset = navigatorVisible ? -LeftPaneDefaultWidth : LeftPaneDefaultWidth
+    dispatch([
+      EditorActions.togglePanel('navigatorPane'),
+      CanvasActions.scrollCanvas({ x: offset, y: 0 } as CanvasVector),
+    ])
+  }, [dispatch, navigatorVisible])
 
   return (
     <SimpleFlexColumn
@@ -290,6 +306,16 @@ export const RightMenu = betterReactMemo('RightMenu', (props: RightMenuProps) =>
               highlightSelected={false}
               icon={<LargerIcons.PreviewPane />}
               onClick={togglePreviewPaneVisible}
+            />
+          </span>
+        </Tooltip>
+        <Tooltip title={'Navigator'} placement={'right'}>
+          <span>
+            <RightMenuTile
+              selected={navigatorVisible}
+              highlightSelected={false}
+              icon={<MenuIcons.Project />}
+              onClick={onClickNavigateTab}
             />
           </span>
         </Tooltip>
